requirements.txt
Django==1.9.4
gevent==1.1.0
greenlet==0.4.9
gunicorn==19.4.5
psycopg2==2.6.1
whitenoise==2.0.6
djangorestframework==3.3.3
